create basic logic for start and end screen
start menu is a separate scene called StartScreen.unity and uses the StartScreenController script to connects to the Game.unity scene.

end menu simply restarts the player to the beginning (stays in the Game.unity scene). Game manager handles and loads the Game.unity scene after restart and the ColliionDetect.cs file is modified to add the restart panel screen.

The following has been done for the current rocks in our segments, but to make this all work for newly created segments, we need to add the Panel object to the restartPanel component in the CollisionDetect script for every obstacle/rock we create from now on.